EPMLABSBRN-0 remove pre-commit and pre-push git hooks
Current pre-commit and pre-push hooks check current user
active branch and fire alert if it's 'master' or 'dev'.
This is not conceptually correct as user works with
his local 'master' and 'dev' copies.

Therefore commit on local 'master' doesn't necesserily means
that it will be pushed to 'origin/master'. Same for 'dev' branch.

Example:
  git push origing master:some-nice-feature

What we really wanted to achieve here is to prevent pushes
to 'origin/master' and 'origin/dev'.

And that should be done on server side only. In our case we need
to make 'dev' and 'master' branches protected in repository settings
on Github.