EPMLABSBRN-311 uncouple AwsConfig and AwsConfig.Conditions classes
Described classes are higly coupled that makes
testing of AwsCloudService harder and requires
mocks on data structures.

In general mocking data structures is anti-pattern
and we should avoid such coding, as it signalizes
us that code have smells and have high coupling.

Instead we need to get rid of such hard links
between classes if they are meant to be used
separately.

As simple solution for now is to change Conditions class
from inner to outer and open for world. In order to decouple
it from parent we need to pass all required params
to constructor or builder class.