EPMLABSBRN-311 update log levels
For all cases that are not actual errors from user perspective
log level was changed from `error` to `warn`
(i.e. user can upload not valid info, or pass not valid param
as part of normal program execution, that's not error
in it's full meaning). This will help to filter only
server side problems in the log.