Remove separate frontend image from building process (#2628)
Right now we have two images for frontend app - one with tls and one without. Right now without tls not used. So to reduce time of building on master branch, I've decided to remove building frontend app without tls, as it is not used anymore after removal of epam instance